pip_requirements.txt
albumentations==1.3.1
black==24.8.0
causal_world==1.2
comet-ml==3.45.1
coverage==7.6.1
colormath
cython==0.29.37
einops==0.6.0
gym==0.21.0
gym_unity==0.27.0
gym[atari, accept-rom-license]==0.21.0
h5py==3.11.0
hydra-core==1.3.2
imageio==2.31.0
imageio-ffmpeg==0.4.8
matplotlib==3.7.1
moviepy==1.0.3
mujoco==2.2.1
mujoco-py==2.1.2.14
numpy==1.23.1
omegaconf==2.3.0
opencv-python==4.7.0.72
pandas==1.5.2
Pillow==9.5.0
portpicker==1.6.0
protobuf==3.20.1
pycocotools==2.0.10
PyOpenGL==3.1.4
pyvirtualdisplay==3.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.10.1
stable-baselines3==1.5.0
spriteworld==1.0.2
tensorboard==2.9.0
tqdm==4.67.1
wandb==0.18.0
